Add support for multiple outputs in C#

+++ csharp/IEnvironment.cs
@@ -5,8 +5,8 @@ namespace TinyNF
     public interface IEnvironment
     {
         // Memory
-        Memory<T> Allocate<T>(nuint size) where T : unmanaged;
-        Memory<T> MapPhysicalMemory<T>(nuint addr, nuint size) where T : unmanaged;
+        Memory<T> Allocate<T>(uint size) where T : unmanaged;
+        Memory<T> MapPhysicalMemory<T>(nuint addr, uint size) where T : unmanaged;
         nuint GetPhysicalAddress<T>(ref T value);
         nuint GetPhysicalAddress<T>(Span<T> span) => GetPhysicalAddress(ref span.GetPinnableReference()); // convenience
